% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalSearch.R
\name{LocalSearch}
\alias{LocalSearch}
\title{Perform Local Hyperparameter Search}
\usage{
LocalSearch(BASIS, Target, nFolds, Epis = "no", foldId = 0, prior = "gaussian")
}
\description{
Perform Local Hyperparameter Search
}
