% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestModel.R
\name{TestModel}
\alias{TestModel}
\title{Test Each Model Using Hyperparameter Combinations}
\usage{
TestModel(
  BASIS,
  Target,
  lambda,
  alpha,
  nFolds,
  foldId = 0,
  Epis = "no",
  prior = "gaussian"
)
}
\description{
Test Each Model Using Hyperparameter Combinations
}
