\name{parEBEN-package}
\alias{parEBEN-package}
\alias{parEBEN}
\docType{package}
\title{Parallel Implementations of the Empirical Bayesian Elastic Net Cross-Validation}

\description{
The parEBEN package provides parallelized cross-validation functionality to the Empirical Bayesian Elastic Net (EBEN) package for handling multicollinearity in generalized linear regression models. By distributing the processing load of the alpha and lambda hyperparameter search across multiple cores of a processor (or multiple machines in a computing cluster), there is a significant reduction in processing time without negatively affecting the output of the EBEN algorithm.

Currently, this package allows for the use of the "doParallel", "doSNOW", and "doMPI" packages, which can parallelize over cores of a single processor or between a series of machines. However, future enhancements will include support for other parallel packages for Hadoop and Spark and possibly GPU-based processing.

}
\details{
\tabular{ll}{
Package: \tab parEBEN\cr
Type: \tab Package\cr
Version: \tab 0.9.9\cr
Date: \tab 2018-11-28\cr
License: \tab Apache 2.0\cr
}

}

\author{Colby T. Ford, Ph.D. - Dept. of Bioinformatics and Genomics, The University of North Carolina at Charlotte}

\references{
Key Algorithms \cr
.	Cai, X., Huang, A., and Xu, S. (2011). Fast empirical Bayesian LASSO for multiple quantitative trait locus mapping. BMC Bioinformatics 12, 211.\cr
.	Huang A, Xu S, Cai X. (2013). Empirical Bayesian LASSO-logistic regression for multiple binary trait locus mapping. BMC genetics  14(1):5. \cr
.	Huang, A., Xu, S., and Cai, X. (2014). Empirical Bayesian elastic net for multiple quantitative trait locus mapping. Heredity 10.1038/hdy.2014.79 \cr\cr
Other Publications: \cr
.	Huang, A., E. Martin, et al. (2014). "Detecting genetic interactions in pathway-based genome-wide association studies." Genet Epidemiol 38(4): 300-309.\cr
.	Huang, A., S. Xu, et al. (2014). "Whole-genome quantitative trait locus mapping reveals major role of epistasis on yield of rice." PLoS ONE 9(1): e87330.\cr
.	Huang, A. (2014). "Sparse model learning for inferring genotype and phenotype associations." Ph.D Dissertation. University of Miami(1186).\cr
.	Wen, J., Ford, C., Janies, D., Shi, X. (2018). "New strategies toward scaling up epistasis analysis on large-scale genomic datasets" ACM Conference on Bioinformatics, Computational Biology, and Health Informatics.
}
\keyword{ package }
